# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZENOVA Platform API",
    "description": "Research-Grade Mental-Health Wellbeing and Conversational Support Platform API",
    "contact": {
      "name": "ZENOVA Engineering & Clinical Team",
      "url": "https://zenova.ai/",
      "email": "support@zenova.ai"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/baseline/{user_id}": {
      "get": {
        "tags": [
          "Personal Baseline"
        ],
        "summary": "Get User Baseline",
        "description": "Retrieve the current personal baseline profile and tracking statistics for a user.",
        "operationId": "get_user_baseline_api_v1_baseline__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get User Baseline Api V1 Baseline  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/baseline/evaluate": {
      "post": {
        "tags": [
          "Personal Baseline"
        ],
        "summary": "Evaluate Observation",
        "description": "Evaluate an inbound multimodal observation against the user's personal baseline.\n\nReturns statistical deviations (z-scores), qualitative interpretations, and\na non-diagnostic clinical disclaimer.",
        "operationId": "evaluate_observation_api_v1_baseline_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultimodalObservation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineDeviationReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/baseline/observation": {
      "post": {
        "tags": [
          "Personal Baseline"
        ],
        "summary": "Record Observation",
        "description": "Record an out-of-band multimodal observation and update longitudinal baseline.",
        "operationId": "record_observation_api_v1_baseline_observation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultimodalObservation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/baseline/reset": {
      "post": {
        "tags": [
          "Personal Baseline"
        ],
        "summary": "Reset User Baseline",
        "description": "Reset baseline records and history for a specific user (for tests or patient consent).",
        "operationId": "reset_user_baseline_api_v1_baseline_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetBaselineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/behavior/observation": {
      "post": {
        "tags": [
          "Behavioral & Passive Sensing"
        ],
        "summary": "Record Observation",
        "description": "Ingest a validated behavioral observation and evaluate multi-feature disruptions.\n\nObservational disclaimer attached. Enforces single-feature non-inference constraint.",
        "operationId": "record_observation_api_v1_behavior_observation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BehavioralObservation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BehavioralAnomalyReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/behavior/ingest-raw": {
      "post": {
        "tags": [
          "Behavioral & Passive Sensing"
        ],
        "summary": "Ingest Raw Device Telemetry",
        "description": "Ingest raw vendor-specific telemetry payload (Apple HealthKit, Fitbit, etc.).",
        "operationId": "ingest_raw_device_telemetry_api_v1_behavior_ingest_raw_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RawWearableIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BehavioralAnomalyReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/behavior/{user_id}/summary": {
      "get": {
        "tags": [
          "Behavioral & Passive Sensing"
        ],
        "summary": "Get User Behavioral Summary",
        "description": "Retrieve longitudinal behavioral baseline summary and tracked distributions for a user.",
        "operationId": "get_user_behavioral_summary_api_v1_behavior__user_id__summary_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get User Behavioral Summary Api V1 Behavior  User Id  Summary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/behavior/{user_id}/clear": {
      "post": {
        "tags": [
          "Behavioral & Passive Sensing"
        ],
        "summary": "Clear User Behavioral Data",
        "description": "Reset and clear behavioral history and profile for a user.",
        "operationId": "clear_user_behavioral_data_api_v1_behavior__user_id__clear_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaselineDeviationReport": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "insufficient_data"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0.0
          },
          "total_observations": {
            "type": "integer",
            "title": "Total Observations",
            "default": 0
          },
          "z_score_valence": {
            "type": "number",
            "title": "Z Score Valence",
            "description": "Deviation of current valence from personal baseline",
            "default": 0.0
          },
          "is_significant_deviation": {
            "type": "boolean",
            "title": "Is Significant Deviation",
            "default": false
          },
          "deviation_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deviation Notes"
          },
          "deviations": {
            "items": {
              "$ref": "#/components/schemas/FeatureDeviation"
            },
            "type": "array",
            "title": "Deviations"
          },
          "metric_deviations": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metric Deviations"
          },
          "deviating_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Deviating Features"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "Statistical deviation from personal baseline; observational and non-diagnostic."
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "BaselineDeviationReport",
        "description": "Evaluation report comparing current observation against personal baseline."
      },
...
      "BehavioralAnomalyReport": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "is_anomaly": {
            "type": "boolean",
            "title": "Is Anomaly",
            "default": false
          },
          "disruption_index": {
            "type": "number",
            "minimum": 0.0,
            "title": "Disruption Index",
            "description": "Composite Behavioral Disruption Index (BDI)",
            "default": 0.0
          },
          "anomalous_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Anomalous Features"
          },
          "feature_z_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Feature Z Scores"
          },
          "observation_count": {
            "type": "integer",
            "title": "Observation Count",
            "default": 0
          },
          "clinical_notes": {
            "type": "string",
            "title": "Clinical Notes",
            "default": "Behavioral patterns within normal individual variance."
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "Passive behavioral signals are observational proxies; do not interpret as clinical diagnoses."
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "BehavioralAnomalyReport",
        "description": "Report evaluating potential multi-feature behavioral disruptions against personal baseline."
      },
...
      "FeatureDeviation": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "current_value": {
            "type": "number",
            "title": "Current Value"
          },
          "baseline_mean": {
            "type": "number",
            "title": "Baseline Mean"
          },
          "baseline_std": {
            "type": "number",
            "title": "Baseline Std"
          },
          "deviation": {
            "type": "number",
            "title": "Deviation",
            "description": "Z-score deviation relative to personal baseline"
          },
          "interpretation": {
            "type": "string",
            "title": "Interpretation",
            "description": "Standardized qualitative deviation categorization"
          }
        },
        "type": "object",
        "required": [
          "feature",
          "current_value",
          "baseline_mean",
          "baseline_std",
          "deviation",
          "interpretation"
        ],
        "title": "FeatureDeviation",
        "description": "Detailed statistical comparison of a specific feature against personal baseline."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MultimodalObservation": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "turn_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn Id"
          },
          "valence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": -1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Valence"
          },
          "arousal": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Arousal"
          },
          "dominant_emotion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dominant Emotion"
          },
          "emotion_confidences": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Emotion Confidences"
          },
          "symptom_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Symptom Scores",
            "description": "Confidence or intensity per identified symptom marker"
          },
          "active_symptom_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Symptom Count"
          },
          "risk_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Level"
          },
          "risk_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Confidence"
          },
          "risk_severity_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 3.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Severity Score"
          },
          "word_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Word Count"
          },
          "sentiment_polarity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Polarity"
          },
          "interaction_interval_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interaction Interval Hours"
          },
          "pitch_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pitch Mean"
          },
          "pitch_std": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pitch Std"
          },
          "speaking_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaking Rate"
          },
          "jitter": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jitter"
          },
          "ema_mood": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Ema Mood"
          },
          "phq9_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 27.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Phq9 Score"
          },
          "gad7_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 21.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Gad7 Score"
          },
          "sleep_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sleep Hours"
          },
          "extra_features": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Extra Features"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "MultimodalObservation",
        "description": "Snapshot observation for an individual user turn or out-of-band event.\n\nGracefully supports missing modalities. None of the modality-specific fields\nare strictly required."
      },
...
      "RawWearableIngestRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "device_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Type",
            "default": "generic_wearable"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "RawWearableIngestRequest"
      },
      "ResetBaselineRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "ResetBaselineRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}